projects moved to page folder, project,project-add,project-edit components changed
