+++ angular-webshop/src/app/app-routing.module.ts
@@ -9,9 +9,9 @@ import { CreateUsersComponent } from "./page/create-users/create-users.component
 import { IndexComponent } from "./page/index/index.component";
 import { OrdersComponent } from "./page/orders/orders.component";
 import { OrderDetailsComponent } from "./page/order-details/order-details.component";
-import { ProjectsComponent } from "./projects/projects.component";
-import { ProjectAddComponent } from "./project-add/project-add.component";
-import { ProjectEditComponent } from "./project-edit/project-edit.component";
+import { ProjectsComponent } from "./page/projects/projects.component";
+import { ProjectAddComponent } from "./page/project-add/project-add.component";
+import { ProjectEditComponent } from "./page/project-edit/project-edit.component";
 import { StatisticsComponent } from "./page/statistics/statistics.component";
 import { UpdateOrdersComponent } from "./page/update-orders/update-orders.component";
 import { UpdateUsersComponent } from "./page/update-users/update-users.component";
