+++ angular-webshop/src/app/app.module.ts
@@ -12,9 +12,9 @@ import { OrdersComponent } from './page/orders/orders.component';
 import { UpdateOrdersComponent } from './page/update-orders/update-orders.component';
 import { UpdateUsersComponent } from './page/update-users/update-users.component';
 import { UsersComponent } from './page/users/users.component';
-import { ProjectAddComponent } from './project-add/project-add.component';
-import { ProjectEditComponent } from './project-edit/project-edit.component'
-import { ProjectsComponent } from './projects/projects.component';
+import { ProjectAddComponent } from './page/project-add/project-add.component';
+import { ProjectEditComponent } from './page/project-edit/project-edit.component'
+import { ProjectsComponent } from './page/projects/projects.component';
 import { BasketComponent } from './page/basket/basket.component';
 import { IndexComponent } from './page/index/index.component';
 import { SearchPipe } from './pipe/search.pipe';
