+++ angular-webshop/src/app/page/project-add/project-add.component.html
@@ -23,13 +23,11 @@
             </div>
             <div class="form-group">
               <label for="shortd">Short description</label>
-              <input type="text" class="form-control" id="shortd" name="project.shortd" [(ngModel)]="newProject.shortd"
-                placeholder="Short description" required>
+              <textarea name="newProject.shortd" id="shortd" class="form-control" [(ngModel)]="newProject.shortd"  rows="5" placeholder="Short description"></textarea>
             </div>
             <div class="form-group">
               <label for="longd">Long description</label>
-              <input type="text" class="form-control" id="longd" name="project.longd" [(ngModel)]="newProject.longd"
-                placeholder="Long description" required>
+              <textarea name="newProject.longd" id="longd" class="form-control" [(ngModel)]="newProject.longd"  rows="10" placeholder="Long description"></textarea>
             </div>
             <div class="form-group">
               <label for="contact">Contact</label>
