+++ angular-webshop/src/app/page/project-edit/project-edit.component.html
@@ -24,13 +24,11 @@
             </div>
             <div class="form-group">
               <label for="shortd">Short description</label>
-              <input type="text" class="form-control" id="shortd" name="project.shortd" [(ngModel)]="project.shortd"
-                placeholder="Short description" required>
+              <textarea name="project.shortd" id="shortd" class="form-control" [(ngModel)]="project.shortd"  rows="5" placeholder="Short description"></textarea>
             </div>
             <div class="form-group">
               <label for="longd">Long description</label>
-              <input type="text" class="form-control" id="longd" name="project.longd" [(ngModel)]="project.longd"
-                placeholder="Long description" required>
+              <textarea name="project.longd" id="longd" class="form-control" [(ngModel)]="project.longd"  rows="10" placeholder="Long description"></textarea>
             </div>
             <div class="form-group">
               <label for="contact">Contact</label>
