+++ angular-webshop/src/app/page/projects/projects.component.css
@@ -3,7 +3,7 @@
   display: block;
   background-position: center;
   z-index: 1;
-  background-color: hsla(0,0%,100%,0.40);
+  background-color: hsla(0, 0%, 100%, 0.40);
   background-blend-mode: overlay;
   background-repeat: no-repeat;
 }
@@ -13,7 +13,8 @@
   display: block;
   background-position: center;
   z-index: 1;
-  background-color: hsla(0,0%,100%,0.80);
+  background-color: hsla(0, 0%, 100%, 0.80);
   background-blend-mode: overlay;
   background-repeat: no-repeat;
 }
+
