+++ angular-webshop/src/app/page/projects/projects.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable, BehaviorSubject } from 'rxjs';
-import { DataService } from '../services/data.service';
+import { DataService } from '../../services/data.service';
 import { Router } from '@angular/router';
-import { Category } from '../model/category';
+import { Category } from '../../model/category';
 
 @Component({
   selector: 'app-projects',
